move logging out of internal package
