+++ logging/func.go
